requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
SQLAlchemy==1.4.46
python-multipart==0.0.5
python-jose==3.3.0
cryptography==39.0.0
passlib==1.7.4
bcrypt==4.0.1
